requirements.txt
Django==1.11
djangorestframework==3.8.2
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2018.3
